+++ packages/@tinacms/graphql/tests/is-title/tina/config.ts
@@ -1,0 +1,20 @@
+import { Schema } from '@tinacms/schema-tools'
+
+export const schema: Schema = {
+  collections: [
+    {
+      name: 'post',
+      path: '',
+      fields: [
+        {
+          type: 'string',
+          name: 'title',
+          isTitle: true,
+          required: true,
+        },
+      ],
+    },
+  ],
+}
+
+export default { schema }